chore: rename cursor to position
