+++ packages/blocks/src/components/format-quick-bar/format-bar-node.ts
@@ -8,7 +8,7 @@ import { html, LitElement } from 'lit';
 import { customElement, property, query, state } from 'lit/decorators.js';
 import { styleMap } from 'lit/directives/style-map.js';
 import { formatConfig } from '../../page-block/utils/const.js';
-import { compareTopAndBottomSpace } from '../../page-block/utils/cursor.js';
+import { compareTopAndBottomSpace } from '../../page-block/utils/position.js';
 import {
   DragDirection,
   getFormat,
