+++ packages/blocks/src/page-block/utils/index.ts
@@ -1,4 +1,4 @@
 export * from './bind-hotkey.js';
 export * from './container-operations.js';
-export { getDragDirection } from './cursor.js';
-export type { DragDirection } from './cursor.js';
+export { getDragDirection } from './position.js';
+export type { DragDirection } from './position.js';
